Add files via upload
the playlist is an excel file which consists of all the songs which a user intend to listen. Letsplay is the main file which is run to automate the youtube.
The project is yet to be tested for all other browsers. It is compatible with Internet Explorer and chrome browser.